some fixes and further refacto TODO: fix the logo on the top left
