+++ makerprint-web/src/components/PrinterPage.js
@@ -2,6 +2,7 @@ import React from 'react';
 import { useParams } from 'react-router-dom';
 import Move from './Move';
 import Status from './Status';
+import FilesList from './FilesList';
 import styles from '../styles/PrinterPage.module.css';
 
 export default function PrinterPage() {
@@ -10,6 +11,7 @@ export default function PrinterPage() {
         <div className="printer-page container">
             <Move />
             <Status />
+            <FilesList />
         </div>
     );
 }
