+++ makerprint-web/src/styles/FilesList.module.css
@@ -1,0 +1,59 @@
+/* .printer_list {
+    display: grid;
+    grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
+    gap: 4px;
+}
+
+.printer_button_wrapper {
+    margin-top: 10px;
+    max-width: 400px;
+    width: 100%;
+    margin: 10px;
+}
+
+.printer_button {
+    display: flex;
+    align-items: center;
+    justify-content: flex-start;
+    width: 100%;
+    padding: 10px 16px;
+    border: none;
+    border-radius: 8px;
+    background-color: #f4f4f4;
+    color: #333333;
+    font-size: 16px;
+    font-weight: bold;
+    cursor: pointer;
+    transition: transform 0.2s;
+    box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
+    position: relative;
+}
+
+.printer_button:hover {
+    transform: scale(1.02);
+}
+
+.printer_button:disabled {
+    opacity: 0.6;
+    cursor: not-allowed;
+}
+
+.printer_icon {
+    height: 48px;
+    width: auto;
+    margin: 0 10px;
+    object-fit: contain;
+}
+
+.printer_name {
+    flex-grow: 1;
+    text-align: left;
+}
+
+.printer_status_indicator {
+    width: 12px;
+    height: 12px;
+    border-radius: 50%;
+    margin-right: 8px;
+    border: 1px solid #33333322;
+} */
